# src/main/resources/static/api.yaml
openapi: "3.0.0"
info:
  title: Airella API
  version: 1.0.0
servers:
  - url: /api
paths:
  /auth/login:
    post:
      operationId: login
      summary: Login user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: |-
            User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: object
                        properties:
                          token:
                            type: string
                          expirationDate:
                            type: string
                      refreshToken:
                        type: string
                      stationRegistrationToken:
                        type: string
        '403':
          description: |-
            Bad credentials
  /auth/refresh-token:
    post:
      operationId: refresh-token
      summary: Refresh user access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: |-
            User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      expirationDate:
                        type: string
        '403':
          description: |-
            Bad refresh token
  /auth/register-user:
    post:
      operationId: register-user
      summary: Register new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: |-
            User created (but not activated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: |-
            This email is already used
  /auth/activate-user:
    get:
      parameters:
        - in: query
          name: activateString
          required: true
          schema:
            type: string
          description: Activation text send to email
      operationId: activate-user
      summary: Activate new user
      responses:
        '200':
          description: |-
            User activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: |-
            This email is already used
  /auth/register-station:
    post:
      operationId: register-station
      summary: Register new station
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                stationRegistrationToken:
                  type: string
                macAddress:
                  type: string
      responses:
        '201':
          description: |-
            New station registered successfully. Returns refreshToken of newly registered station. Id of station is returned in Location header.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        description: "Id of newly registered station"
                      refreshToken:
                        type: string
                        description: "Token of newly registered station"
        '400':
          description: |-
            Bad stationRegistrationToken
        '403':
          description: |-
            Bad refresh token
  /stations:
    get:
      operationId: get-stations
      summary: Lists stations
      responses:
        '200':
          description: |-
            Returns all stations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/station-brief'
  /stations/{stationId}:
    get:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
        - in: query
          name: timespan
          required: false
          schema:
            type: string
            format: date-time
          description: Timespan of queried measurements in ISO 8601 format in UTC (must be 'Z' at the end)
          example: 2020-04-02T21:37:00/2020-04-05T22:48:00Z
        - in: query
          name: interval
          required: false
          schema:
            type: string
            format: date-time
          description: Interval between queried measurements in ISO 8601 format
          example: PT1H
        - in: query
          name: sensors
          required: false
          schema:
            type: string
          description: Filter for sensors - split by comma
          example: pm10,pm25
        - in: query
          name: strategy
          required: false
          schema:
            type: string
          description: Strategy of agregating results - avg, all or latest
          example: avg
      operationId: get-station
      summary: Returns information about one station
      responses:
        '200':
          description: |-
            Returns information about one station, including sensor ones.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/station'
    delete:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: delete-station
      summary: Deletes station
      responses:
        '200':
          description: |-
            Deletes station, including sensor ones.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /map/at:
    get:
      parameters:
        - in: query
          name: latitude
          required: true
          schema:
            type: number
          description: Latitude of queried point
          example: 50.0680472
        - in: query
          name: longitude
          required: true
          schema:
            type: number
          description: Longitude of queried point
          example: 19.9104095
      operationId: map-at
      summary: Returns information about selected point (interpolated)
      responses:
        '200':
          description: |-
            Returns information about selected point.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/virtual-station'
  /stations/{stationId}/name:
    put:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: update-station-name
      summary: Updates station name
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/station-name'
      responses:
        '200':
          description: |-
            Updates station name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /stations/{stationId}/address:
    put:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: update-station-address
      summary: Updates station address
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/address'
      responses:
        '200':
          description: |-
            Updates station address.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /stations/{stationId}/location:
    put:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: update-station-location
      summary: Updates station location
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/location'
      responses:
        '200':
          description: |-
            Updates station location.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /stations/{stationId}/sensors:
    post:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: add-sensor
      summary: Adds sensor to station
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                id:
                  type: string
      responses:
        '201':
          description: |-
            Adds new sensor. Sensor URI is returned in Location header.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
  /stations/{stationId}/sensors/{sensorId}/measurements:
    post:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
        - in: path
          name: sensorId
          required: true
          schema:
            type: integer
            minimum: 1
          description: Sensor id
      operationId: add-measurement
      summary: Adds measurement to sensor
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: number
      responses:
        '200':
          description: |-
            Added new measurement with current date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /search/map:
    get:
      parameters:
        - in: query
          name: latitude
          required: true
          schema:
            type: number
          description: Latitude of searched area
          example: 50.0680472
        - in: query
          name: longitude
          required: true
          schema:
            type: number
          description: Longitude of searched area
          example: 19.9104095
        - in: query
          name: radius
          required: true
          schema:
            type: number
          description: Radius of searched area in meters
          example: 10000
      operationId: search-map
      summary: Searchs for stations in selected area
      responses:
        '200':
          description: |-
            Successfully searched for stations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/station-brief'
  /user/stations:
    get:
      operationId: get-user-stations
      summary: Lists user stations
      responses:
        '200':
          description: |-
            Returns user stations. User needs to be logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/station-brief'
  /stations/{stationId}/statistics/{statisticId}/values:
    post:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
        - in: path
          name: statisticId
          required: true
          schema:
            type: string
            minimum: 1
          description: Statistic id
      operationId: add-to-statistic
      summary: Add statistic value to statistic
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: object
                  description: value of new statistic value
      responses:
        '200':
          description: |-
            Added new element to statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /stations/{stationId}/statistics:
    post:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: add-statistic
      summary: Adds statistic to station
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                statisticPrivacyMode:
                  type: string
                  description: public or private
                type:
                  type: string
                  description: public or private
                id:
                  type: string
                  description: one_string_value, one_double_value, multi_string_value, multi_double_aggregatable_value, multi_double_value
      responses:
        '201':
          description: |-
            New statistic registered successfully. Id of statistic is returned in Location header.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        description: "Id of newly registered statistic"
    get:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
      operationId: get-statistics
      summary: Returns information about statistics for station
      responses:
        '200':
          description: |-
            Returns statistics about one station.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/statistics'
  /stations/{stationId}/statistics/{statisticId}:
    delete:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
        - in: path
          name: statisticId
          required: true
          schema:
            type: string
            minimum: 1
          description: Statistic id
      operationId: remove-statistic
      summary: Remove statistic from station
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                success:
                  type: boolean
      responses:
        '200':
          description: |-
            Deletes statistic.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    get:
      parameters:
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            minimum: 1
          description: Station id
        - in: path
          name: statisticId
          required: true
          schema:
            type: string
            minimum: 1
          description: Statistic id
        - in: query
          name: timespan
          required: false
          schema:
            type: string
            format: date-time
          description: Timespan of queried measurements in ISO 8601 format in UTC (must be 'Z' at the end)
          example: 2020-04-02T21:37:00/2020-04-05T22:48:00Z
        - in: query
          name: interval
          required: false
          schema:
            type: string
            format: date-time
          description: Interval between queried measurements in ISO 8601 format
          example: PT1H
        - in: query
          name: strategy
          required: false
          schema:
            type: string
          description: Strategy of agregating results - for example latest
          example: latest
      operationId: get-statistic
      summary: Returns information about one statistic
      responses:
        '200':
          description: |-
            Returns information about one statistic.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/statistic'
components:
  schemas:
    sensor-brief:
      type: object
      properties:
        type:
          type: string
        value:
          type: number
    sensor:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/measurement'
        status:
          type: number
    station-brief:
      type: array
      items:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
          address:
            $ref: '#/components/schemas/address'
          location:
            $ref: '#/components/schemas/location'
          aqi:
            type: number
    station:
      type: array
      items:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
          address:
            $ref: '#/components/schemas/address'
          location:
            $ref: '#/components/schemas/location'
          aqi:
            type: number
          sensors:
            type: array
            items:
              $ref: '#/components/schemas/sensor'
    virtual-station:
      type: array
      items:
        type: object
        properties:
          aqi:
            type: number
          sensors:
            type: array
            items:
              $ref: '#/components/schemas/sensor'
    address:
      type: object
      properties:
        country:
          type: string
        city:
          type: string
        street:
          type: string
        number:
          type: string
    location:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
    station-name:
      type: object
      properties:
        name:
          type: string
    measurement:
      type: object
      properties:
        timespan:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        timestamp:
          type: string
          format: date-time
        value:
          type: number
    statistic:
      type: object
      properties:
        timespan:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        timestamp:
          type: string
          format: date-time
        value:
          type: object
    statistics:
      type: object
      properties:
        statistics:
          type: array
          items:
            $ref: '#/components/schemas/statistic'